Initial project commit
App.js manages app state with a router and switch. Added /login and
/register paths that route to the login and register components
respectively. Added PrivateRoute component that takes care of
conditionally redirecting to another route if a condition is not met. A
private route is in place for the / path, which redirects to login if
the user is not authenticated. A (not exact) route is defined for the
/ path, which redirects back to the / path, which is done to redirect
paths that don't match a defined route back to /.

Login component takes in user email and password and is used to
authenticate with the api server.

Register component takes in a user email, name, and password and is used
to register with the api server.

Modal component is used to display modal views. LoadingModal component displays a
modal with a title and loading indicator. MessageModal displays a modal
with a title and message as well as a confirmation button that triggers
the onConfirm function passed as a prop.