Change modal propTypes and fix modal overlay
Changed the modal (and LoadingModal and MessageModal) propTypes so that
visible is not required.
Changed the modal overlay css class to add a z-index of 100 so that it properly
overlays other components. Also changed it to using
top/left/right/bottom: 0 instead of width/height: 100% to properly
stretch and ignore parent margins and padding.