Add TransactionCreator component
TransactionCreator component was created for users to be able to
send/request coins to other users. It takes in the list of the user's
coins and a function to call when the transaction is submitted as props.
If the inputs are acceptable (pass validation) when the submit button is
pressed, the submit function prop is called with an object containing
the chargeValue (true for charge, false for send), receiver UUID to send
the transaction to, amount to create the transaction with, coin to
create the transaction with, and an optional message to send with the
transaction.